requirements.txt
kagglehub
pandas
matplotlib
openai
arize-phoenix
dotenv
streamlit
langchain
langchain_openai
langchain_community
langchain_chroma
openinference-instrumentation-langchain